+++ src/main/java/bridge/BridgeMaker.java
@@ -1,4 +1,4 @@
-package bridge.domain.maker;
+package bridge;
 
 import java.util.ArrayList;
 import java.util.List;
