+++ src/main/java/bridge/controller/Controller.java
@@ -9,8 +9,8 @@ import java.util.List;
 import bridge.constant.Direction;
 import bridge.constant.GameCommand;
 import bridge.domain.BridgeGame;
-import bridge.domain.maker.BridgeMaker;
-import bridge.domain.maker.BridgeRandomNumberGenerator;
+import bridge.BridgeMaker;
+import bridge.BridgeRandomNumberGenerator;
 import bridge.dto.TrialResult;
 import bridge.view.InputView;
 import bridge.view.OutputView;
